edit table for maintenance

+++ src/lib/chip.svelte
@@ -3,8 +3,9 @@
     import { MDCChipSet } from '@material/chips';
     import { onMount, createEventDispatcher } from 'svelte';
 
-    // export let chip = {}
     export let chips = []
+    export let action = 'remove'
+    export let type = 'filled'
 
     let chipset
     let chipsetComponent
@@ -16,32 +17,41 @@
     })
 
 </script>
-<span bind:this={chipsetComponent} class="mdc-evolution-chip-set" role="listbox" aria-orientation="horizontal">
+<span 
+  bind:this={chipsetComponent} 
+  class="mdc-evolution-chip-set" 
+  role="listbox" 
+  aria-orientation="horizontal">
     <span class="mdc-evolution-chip-set__chips" role="presentation">
       {#each chips as chip }
-        <span class="mdc-evolution-chip" role="presentation" id={chip.id} >
+        <span class="mdc-evolution-chip" class:mdc-chip--outlined={type === 'outlined'} role="presentation" id={chip.id} >
           <span class="mdc-evolution-chip__cell mdc-evolution-chip__cell--primary" role="gridcell">
-            <button class="mdc-evolution-chip__action mdc-evolution-chip__action--primary" type="button" tabindex="0" data-mdc-deletable="true" on:click>
-              <span class="mdc-evolution-chip__ripple mdc-evolution-chip__ripple--primary"></span>
-              <span class="mdc-evolution-chip__text-label">{chip.label}</span>
-            </button>
+            <!-- {#if chip?.action } -->
+              <button class="mdc-evolution-chip__action mdc-evolution-chip__action--primary" type="button" tabindex="0" data-mdc-deletable="true" on:click>
+                <span class="mdc-evolution-chip__ripple mdc-evolution-chip__ripple--primary"></span>
+                <span class="mdc-evolution-chip__text-label">{chip.label}</span>
+              </button>
+            <!-- {/if} -->
           </span>
-          <button 
-            class="mdc-evolution-chip__action mdc-evolution-chip__action--trailing" 
-            type="button" 
-            tabindex="-1" 
-            data-mdc-deletable="true" 
-            aria-hidden="true"
-            on:click={ () => {
-              // Eliminar el chip de chips
-              dispatch('deleteFilter', chip)
-            } }
-          >
-            <span class="mdc-evolution-chip__ripple mdc-evolution-chip__ripple--trailing"></span>
-            <span class="mdc-evolution-chip__icon mdc-evolution-chip__icon--trailing material-symbols-rounded">close</span>
-            <!-- <span class="material-symbols-rounded">close</span> -->
+          {#if action != ''}
+            <button 
+              class="mdc-evolution-chip__action mdc-evolution-chip__action--trailing" 
+              type="button" 
+              tabindex="-1" 
+              data-mdc-deletable="true" 
+              aria-hidden="true"
+              on:click={ () => {
+                // Eliminar el chip de chips
+                if (action === 'remove') dispatch('deleteFilter', chip)
+                else dispatch(chip.dispatch, chip)
+              } }
+            >
+              <span class="mdc-evolution-chip__ripple mdc-evolution-chip__ripple--trailing"></span>
+              <span class="mdc-evolution-chip__icon mdc-evolution-chip__icon--trailing material-symbols-rounded">{ action == 'remove' ? 'close' : chip.icon }</span>
+              <!-- <span class="material-symbols-rounded">close</span> -->
 
-          </button>
+            </button>
+          {/if}
         </span>
       {/each}
     </span>
@@ -51,4 +61,21 @@
     .mdc-evolution-chip__action {
       margin-right: 8px;
     }
+
+    .mdc-chip--outlined {
+      border: 1px solid rgba(0, 0, 0, 0.12);
+      background-color: transparent;
+    }
+
+    .red {
+      border-color: red;
+    }
+
+    .green {
+      border-color: green;
+    }
+
+    .yellow {
+      border-color: yellow;}
+
 </style>
